絞り込み '優先度' 実装
