+++ pages/components/Search.jsx
@@ -1,6 +1,31 @@
 import { Box, Container, Input, Select, Text } from "@chakra-ui/react";
+import { useEffect, useState } from "react";
+import { useRecoilValue } from "recoil";
+import { todosState } from "../atoms/atom";
 
-const Search = () => {
+const Search = ({ setPriority }) => {
+  // const handleSearchPriority = (e) => {
+  //   setPriority(e.target.value);
+  //   console.log(e)
+
+  //   setSearchTodos(todos.filter((todo) => {
+  //     return todo.priority === priority;
+  //   }));
+  // }
+
+  // useEffect(() => {
+  //   const newTodos = todos.filter((todo) => {
+  //     return todo.priority === priority
+  //   })
+
+  //   setSearchTodos(newTodos);
+
+  // }, [priority, todos])
+
+  // useEffect(() => {
+  //   console.log(searchTodos)
+  // }, [searchTodos])
+
   return (
     <Container m='0' mt='10' p='3' bg='gray.100' borderRadius='10px'  display='flex' justifyContent='space-around' alignItems='center'>
       <Box>
@@ -19,7 +44,11 @@ const Search = () => {
 
       <Box>
         <Text textAlign='center'>優先度</Text>
-        <Select placeholder='すべて'>
+        <Select 
+          defaultValue={'すべて'}
+          onChange={(e) => setPriority(e.target.value)}
+        >
+          <option value='すべて'>すべて</option>
           <option value='低'>低</option>
           <option value='中'>中</option>
           <option value='高'>高</option>
